fix(FieldPreset\Inline): create a foreign match field for the child table if needed
To make sure a table can create multiple inline fields pointing to the same child table,
a the local field name must be stored in the foreign table. v11 will therefore automatically
create two additional fields on the child table, one for storing the name of the field,
the other for storing the name of the table. This allows the script to determine the source
of the parent more reliably. Sadly this breaks existing configurations. To prevent issues,
the feature is only active if the matching feature toggle is enabled.

Also an upgrade wizard is provided to migrate existing records to match the new requirements.