fix(ExtConfig): ensure config state consistency when TCA is loaded multiple times
This happens in the install-tool or via cli commands. The old implementation lead to side-effects when some config handlers were executed multiple times on the same, persistent state object. The end result were option arrays with hundreds or thousands of entries, especially in the (icon) tables.

From now on the state without the TCA extensions is stored next to the one with the TCA extensions. When the TCA files are loaded, the clean state is restored and the extensions are applied anew.