fix(EventBus): fix double registration of listeners in install tool
The install-tool extension has a functionality called the "LateBootService"
which allows the di container to be recreated for some reason.
If that is happening (like in upgrade wizards) some issues arose,
like event listeners being re-registered multiple times.
To prevent that issue, we check if the current stage is already instantiated
and handle the creation differently if we re-instantiate an existing container.